model presenter method for providing resources to cells tweaked.
This implementation appears to be better than the old implementation
that required an explcit declaration of every resource that a cell
might require with no resuse elsewhere.

In the new model relies on the assumption that the cell already
has the resource it needs and the only information needed from the
outside it to identify that cell and tell it what has changed in the
resource to enable updating the cell ui.